update ink server
