+++ .gitignore
@@ -2,3 +2,4 @@ ink
 public
 release
 debian
+npm-debug.log
