+++ editor/assets/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <title>纸小墨 | 编辑器</title>
+    </head>
+    <body>
+        <div id="root"></div>
+    </body>
+    <script src="bundle.js"></script>
+</html>
