--- editor/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="UTF-8">
-        <title>纸小墨 | 编辑器</title>
-    </head>
-    <body>
-        <div id="root"></div>
-    </body>
-    <script src="/static/bundle.js"></script>
-</html>
