+++ editor/source/styles/index.less
@@ -127,7 +127,7 @@
     right: 0;
     z-index: 1;
     text-align: center;
-    margin-top: 100px;
+    margin-top: 60px;
     .title {
         font-size: 32px;
         font-weight: bold;
